avoid various gcc warnings
git-svn-id: https://svn.salilab.org/imp/trunk@5691 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ kernel/src/Key.cpp
@@ -18,7 +18,8 @@ KeyData::KeyData(): heuristic_(heuristic_value){}
 
 void KeyData::assert_is_initialized() const
 {
-  IMP_INTERNAL_CHECK(heuristic_== heuristic_value,
+  IMP_INTERNAL_CHECK(static_cast<int>(heuristic_)
+                     == static_cast<int>(heuristic_value),
              "Uninitialized KeyData. Do not initialize Keys statically.");
 }
